Changed the logging stuff to have a container class for...
Author: dustin
Date: 1999/09/30 06:53:02

Changed the logging stuff to have a container class for log entries,
and then extend it for the Image log.  This gives the flexibility to try more
log things in the future.

Changed the name of the log table from ``log'' to ``photo_log''

imported from: dustin@spy.net--projects/snippets--java--1.0--patch-103