Some LRU bug fixes.
It turns out, the bug is not in the LRU library, but in the Linkedlist
module. It used structural comparison (el = it) instead of physical
comparison (el == it) and probably hit some GC-related timing issue in
a newer OCaml version (I don't believe it did not ever work for you,
so I presume it worked on some earlier OCaml version).

However, I've also found a problem in it's to_list code: it duplicated
the entry in the returned list instead of showing it only once. I also
took the chance to make it sorted descenging in the order of usage
(first item in the list returned by to_list is the most recently used
one).