Updated frogger cam mount for installation.
I actually think it's *slightly* off in a couple of ways, but it'll
work.

Could be a little narrower all around. The standoffs are just barely
too far out.  The actual frogger standoffs are themselves narrower
than the nylon ones I like to use.

Also, the actual part the holds the camera feels looser than
necessary.  This is likely due to it not being a single cylinder
shape.