Disable the backlight most of the time.
It's enabled for 15s after data changes, or when someone touches it to
see what's going on.

I should probably specify when to turn it off instead of when
something interesting happened.  This would allow errors to display
longer than data.

The light pad is jumped to pin 21 via a 0Ω resistor.  Convenient fit,
and still allows me to upload code.

I think I'd rather do something fancy with pwm, but I've not toyed
with that API much yet.