Watchdog updates.
I saw a display hanging while trying to connect to the network.  Not
sure how long it was like that, but if it can't get data quickly, it
should just go back to sleep and try later.

Cool thing, though, is I had a major network outage after this and it
did all the right things.